requirements.txt
functions-framework==3.*
requests==2.28.1
python-dotenv==0.21.0
geoip2==4.6.0